Unbind errors collection from the policy
In the v1.0 and before the collection of errors referred
to the policy that this errors originated from. The only usage
of this policy was extraction of a proper scope for error translation.

From now and on we build collections of errors with an explicit
scope instead of referring to the scope via policy. This makes
the collection more generic and reusable outside of the policy context.
The scope is added to every single error in the collection as well.

By default the scope is set to ["tram-policy", "errors"].

This commit is backward incompatible for several reasons:
- unbinding [Tram::Policy::Errors] from a policy
- changing the signature of [Tram::Policy::Errors#initialize]
- changing the signature of [Tram::Policy::Error#initialize]